+++ src/cli/main.ts
@@ -1,0 +1,22 @@
+#!/usr/bin/env node
+import 'reflect-metadata';
+import { readFile } from 'node:fs/promises';
+import { Command } from 'commander';
+import { getPackageJsonPath } from '../utils/index.js';
+import { importCommand } from './commands/import.js';
+import { generateCommand } from './commands/generate.js';
+
+const packageUrl = getPackageJsonPath();
+const packageJsonContent = JSON.parse(await readFile(packageUrl, 'utf-8'));
+
+const program = new Command();
+
+program
+  .name('node-tsv-parser')
+  .description('Parse .tsv files with node')
+  .version(packageJsonContent.version, '-v, --version');
+
+program.addCommand(importCommand);
+program.addCommand(generateCommand);
+
+program.parse(process.argv);
